Update README.md
The research project Diagnostic Assessment & Achievement of College Skills (DAACS) is supported by a grant from the U.S. Department of Education. This repository contains Oxana Rosca's R-scripts on organization and analyses of the data collected by the DAACS application. The data are not publicly available due to their containing information that could compromise the privacy of research participants.